Refactor heartbeat event handling to suppress workflow notices for consolidation carryover summaries
- Removed sending of workflow notices for consolidation strategy learning updates and dream state updates in `postprocess.rs`, replacing them with debug logs.
- Updated `fetch_models` function in `helpers.rs` to apply attribution headers correctly.
- Modified tests in `part3.rs` to check for correct request texts and added assertions for attribution headers.
- Enhanced adaptive carryover summary in `tests.rs` to include source kind and provenance reference.
- Introduced `adaptive_carryover_source_kind` function in `provenance.rs` to determine source kind based on event type.
- Updated carryover summary structure to include inspection tool in `provenance.rs`.
- Adjusted tests in `consolidation.rs` to verify inspection tool presence in carryover summaries.
- Altered heartbeat test in `heartbeat.rs` to ensure no workflow notices are emitted during structured heartbeat processing.
- Changed default debate configuration to enabled in `config_core.rs` and updated related tests.
- Improved embedding queue logic to handle deletion of indexed messages correctly in `embedding_queue.rs`.
- Enhanced vector index handling for batch upserts and deletions in `vector_index.rs`.
- Added context invalidation logic in TUI event handling to ensure proper context management during auto-compaction.
